Add files via upload
Added an option to skip the curve entirely and to plot CI curves instead of a ribbon. Also, in the case when a ribbon is specified but the upper and lower CI curves have different lengths, I set up the plot to put out the curves and ignore the ribbon call.